[hist] Count bins as std::uint64_t
This is semantically more correct because for sparse histograms, not
all bins have to be allocated in memory. For RRegularAxis, it also
makes the member definition portable across platforms with different
widths of std::size_t.

+++ hist/histv7/test/hist_index.cxx
@@ -1,5 +1,6 @@
 #include "hist_test.hxx"
 
+#include <cstdint>
 #include <iterator>
 #include <vector>
 
@@ -56,7 +57,7 @@ TEST(RBinIndex, Plus)
    }
 
    // Matches RBinIndex::UnderflowIndex
-   static constexpr std::size_t UnderflowIndex = -3;
+   static constexpr std::uint64_t UnderflowIndex = -3;
    EXPECT_TRUE((RBinIndex(0) + UnderflowIndex).IsInvalid());
    EXPECT_TRUE((RBinIndex(3) + UnderflowIndex).IsInvalid());
 }
